adding some cli and comments added tests

+++ cpp/example_code/dynamodb/dynamodb_getting_started_scenario.cpp
@@ -42,10 +42,13 @@ namespace AwsDoc {
                                        std::function<bool(Aws::String)> test = [](
                                                const Aws::String &) -> bool { return true; });
 
-        int askQuestionForInt(const Aws::String &string);
+        int askQuestionForInt(const std::string &string);
 
         float
         askQuestionForFloatRange(const Aws::String &string, float low, float high);
+
+        int askQuestionForIntRange(const Aws::String &string, int low,
+                                                     int high);
 
         static bool waitTableActive(const Aws::String &tableName,
                                     const Aws::Client::ClientConfiguration &clientConfiguration);
@@ -58,22 +61,25 @@ namespace AwsDoc {
         void printMovieInfo(
                 const Aws::Map<Aws::String, Aws::DynamoDB::Model::AttributeValue> &movieMap);
 
-        static int inflateZip(FILE *source, FILE *dest);
+        static int deflateZip(FILE *source, FILE *dest);
 
     } //  namespace DynamoDB
 } // namespace AwsDoc
 
 /*
  *
- * 1. Create a table with partition: year (N) and sort: title (S) (CreateTable)
+ * 1. Create a table with partition: year (N) and sort: title (S). (CreateTable)
  * 2. Add a new movie. (PutItem)
- * 3. Update the rating and plot of the movie by using an update expression. (UpdateItem with UpdateExpression + ExpressionAttributeValues args)
- * 4. Put movies in the table from moviedata.json--download it from the DynamoDB guide. OK to include manual download and unzip steps in your example README. This is a large file so limit the number of movies to 250 or so. (BatchWriteItem)
- * 5. Get a movie by Key (partition + sort) (GetItem)
- * 6. Use Query with a key condition expression to return all movies released in a given year. (Query + KeyConditionExpression arg)
- * 7. Use Scan to return movies released within a range of years. Show how to paginate data using ExclusiveStartKey. (Scan + FilterExpression)
+ * 3. Update the rating and plot of the movie by using an update expression.
+ *    (UpdateItem with UpdateExpression + ExpressionAttributeValues args)
+ * 4. Put movies in the table from moviedata.json--download it from the DynamoDB guide. (BatchWriteItem)
+ * 5. Get a movie by Key. (partition + sort) (GetItem)
+ * 6. Use Query with a key condition expression to return all movies released in a given
+ *    year. (Query + KeyConditionExpression arg)
+ * 7. Use Scan to return movies released within a range of years. Show how to
+ *    paginate data using ExclusiveStartKey. (Scan + FilterExpression)
  * 8. Delete a movie. (DeleteItem)
- * 9.Delete the table. (DeleteTable)
+ * 9. Delete the table. (DeleteTable)
  */
 
 bool AwsDoc::DynamoDB::dynamodbGettingStartedScenario(
@@ -86,7 +92,7 @@ bool AwsDoc::DynamoDB::dynamodbGettingStartedScenario(
 
     bool movieTableAlreadyExisted = false;
 
-    // 1. Create a table with partition: year (N) and sort: title (S) (CreateTable)
+    // 1. Create a table with partition: year (N) and sort: title (S).
     {
         Aws::DynamoDB::Model::CreateTableRequest request;
 
@@ -136,6 +142,7 @@ bool AwsDoc::DynamoDB::dynamodbGettingStartedScenario(
         }
     }
 
+    // Wait for table to become active.
     if (!movieTableAlreadyExisted) {
         std::cout << "Waiting for table '" << MOVIE_TABLE_NAME
                   << "' to become active...." << std::endl;
@@ -147,7 +154,7 @@ bool AwsDoc::DynamoDB::dynamodbGettingStartedScenario(
                   << std::endl;
     }
 
-    // 2. Add a new movie. (PutItem)
+    // 2. Add a new movie.
     Aws::String title;
     float rating;
     int year;
@@ -196,7 +203,6 @@ bool AwsDoc::DynamoDB::dynamodbGettingStartedScenario(
               << std::endl;
 
     // 3. Update the rating and plot of the movie by using an update expression.
-    // (UpdateItem with UpdateExpression + ExpressionAttributeValues args)
     {
         rating = askQuestionForFloatRange(
                 Aws::String("\nLet's update your movie.\nYou rated it  ") +
@@ -234,8 +240,9 @@ bool AwsDoc::DynamoDB::dynamodbGettingStartedScenario(
 
     std::cout << "\nUpdated '" << title << "' with new attributes:" << std::endl;
 
+    // 4. Put 250 movies in the table from moviedata.json
     if (!movieTableAlreadyExisted) {
-        // 4. Put 250 movies in the table from moviedata.json (BatchWriteItem)
+        std::cout << "Adding movies from a json file to the database." << std::endl;
         const size_t MAX_SIZE_FOR_BATCH_WRITE = 25;
         const size_t MOVIES_TO_WRITE = 10 * MAX_SIZE_FOR_BATCH_WRITE;
         Aws::String jsonString = getMovieJSON();
@@ -244,8 +251,9 @@ bool AwsDoc::DynamoDB::dynamodbGettingStartedScenario(
             Aws::Utils::Array<Aws::Utils::Json::JsonView> movieJsons = json.View().AsArray();
             Aws::Vector<Aws::DynamoDB::Model::WriteRequest> writeRequests;
 
-            // Movies are grouped by year. To get a cross-section of years, get every
-            // "increment" movie.
+            // Movies are grouped by year in the json file. To add movies with a
+            // cross-section of years, use an appropriate increment value for iterating
+            // through the database.
             size_t increment = movieJsons.GetLength() / MOVIES_TO_WRITE;
             for (size_t i = 0; i < movieJsons.GetLength(); i += increment) {
                 writeRequests.push_back(Aws::DynamoDB::Model::WriteRequest());
@@ -279,23 +287,22 @@ bool AwsDoc::DynamoDB::dynamodbGettingStartedScenario(
 
     std::cout << std::setfill('*') << std::setw(88) << " " << std::endl;
 
-    // 5. Get a movie by Key (partition + sort)
-    //    (GetItem)
+    // 5. Get a movie by Key (partition + sort).
     {
-        Aws::String titleToGet("Star Trek: First Contact");
-        //     Aws::String answer = askQuestion(Aws::String("Let's move on...do you want to get info about '" + titleToGet + "'? (y/n) "));
-        if (false) //(answer == "y")
+        Aws::String titleToGet("King Kong");
+        Aws::String answer = askQuestion(Aws::String("Let's move on...Wou to get info about '" + titleToGet + "'? (y/n) "));
+        if (answer == "y")
         {
             Aws::DynamoDB::Model::GetItemRequest request;
             request.SetTableName(MOVIE_TABLE_NAME);
             request.AddKey(TITLE_KEY,
                            Aws::DynamoDB::Model::AttributeValue().SetS(titleToGet));
-            request.AddKey(YEAR_KEY, Aws::DynamoDB::Model::AttributeValue().SetN(2013));
+            request.AddKey(YEAR_KEY, Aws::DynamoDB::Model::AttributeValue().SetN(1933));
 
             const Aws::DynamoDB::Model::GetItemOutcome &result = dynamoClient.GetItem(
                     request);
             if (!result.IsSuccess()) {
-                std::cout << "Error " << result.GetError().GetMessage();
+                std::cerr << "Error " << result.GetError().GetMessage();
                 deleteDynamoTable(MOVIE_TABLE_NAME, clientConfiguration);
                 return false;
             }
@@ -306,7 +313,7 @@ bool AwsDoc::DynamoDB::dynamodbGettingStartedScenario(
                     printMovieInfo(item);
                 }
                 else {
-                    std::cout << "The movie was not found in the database."
+                    std::cout << "\nThe movie was not found in the database."
                               << std::endl;
                 }
             }
@@ -314,7 +321,8 @@ bool AwsDoc::DynamoDB::dynamodbGettingStartedScenario(
     }
 
     // 6. Use Query with a key condition expression to return all movies released in a given year.
-    //    (Query + KeyConditionExpression arg)
+    Aws::String doAgain = "n";
+    do
     {
         Aws::DynamoDB::Model::QueryRequest req;
 
@@ -325,10 +333,12 @@ bool AwsDoc::DynamoDB::dynamodbGettingStartedScenario(
         req.SetKeyConditionExpression("#dynobase_year = :valueToMatch");
         req.SetExpressionAttributeNames({{"#dynobase_year", YEAR_KEY}});
 
-        // Set Expression AttributeValues
+        int yearToMatch = askQuestionForIntRange("\nLet's get a list of movies released in"
+                                            " a given year. Enter a year between 1972 and 2018 ",
+                                            1972, 2018);
         Aws::Map<Aws::String, Aws::DynamoDB::Model::AttributeValue> attributeValues;
         attributeValues.emplace(":valueToMatch",
-                                Aws::DynamoDB::Model::AttributeValue().SetN(2013));
+                                Aws::DynamoDB::Model::AttributeValue().SetN(yearToMatch));
         req.SetExpressionAttributeValues(attributeValues);
 
         // Perform Query operation
@@ -337,32 +347,35 @@ bool AwsDoc::DynamoDB::dynamodbGettingStartedScenario(
             // Reference the retrieved items
             const Aws::Vector<Aws::Map<Aws::String, Aws::DynamoDB::Model::AttributeValue>> &items = result.GetResult().GetItems();
             if (items.size() > 0) {
-                std::cout << "Number of items retrieved from Query: " << items.size()
-                          << std::endl;
-                //Iterate each item and print
+                std::cout << "\nThere were " << items.size() << " movies in the database from "
+                    << yearToMatch << "." << std::endl;
+                 //Iterate each item and print
                 for (const auto &item: items) {
-                    std::cout
-                            << "******************************************************"
-                            << std::endl;
-                    // Output each retrieved field and its value
-                    for (const auto &i: item)
-                        std::cout << i.first << ": " << i.second.GetS() << std::endl;
+                    printMovieInfo(item);
                 }
+                doAgain = "n";
             }
             else {
-                std::cout << "No item found in table: " << MOVIE_TABLE_NAME
+                std::cout << "\nNo movies from " << yearToMatch << " were found in the database"
                           << std::endl;
+                doAgain = askQuestion(Aws::String("Try another year? (y/n) "));
             }
         }
         else {
-            std::cout << "Failed to Query items: " << result.GetError().GetMessage();
+            std::cerr << "Failed to Query items: " << result.GetError().GetMessage();
         }
-    }
+
+
+    } while (doAgain == "y");
 
     //  7. Use Scan to return movies released within a range of years.
     //     Show how to paginate data using ExclusiveStartKey. (Scan + FilterExpression)
-
     {
+        int startYear = askQuestionForIntRange("\nNow let's scan a range of years "
+                                               "for movies in the database. Enter a start year: ",
+                                               1972, 2018);
+        int endYear = askQuestionForIntRange("\nEnter an end year: ",
+                                             startYear, 2018);
         Aws::Map<Aws::String, Aws::DynamoDB::Model::AttributeValue> exclusiveStartKey;
         do {
             Aws::DynamoDB::Model::ScanRequest scanRequest;
@@ -373,9 +386,9 @@ bool AwsDoc::DynamoDB::dynamodbGettingStartedScenario(
 
             Aws::Map<Aws::String, Aws::DynamoDB::Model::AttributeValue> attributeValues;
             attributeValues.emplace(":startYear",
-                                    Aws::DynamoDB::Model::AttributeValue().SetN(2009));
+                                    Aws::DynamoDB::Model::AttributeValue().SetN(startYear));
             attributeValues.emplace(":endYear",
-                                    Aws::DynamoDB::Model::AttributeValue().SetN(2011));
+                                    Aws::DynamoDB::Model::AttributeValue().SetN(endYear));
             scanRequest.SetExpressionAttributeValues(attributeValues);
 
             if (!exclusiveStartKey.empty()) {
@@ -387,18 +400,27 @@ bool AwsDoc::DynamoDB::dynamodbGettingStartedScenario(
             if (result.IsSuccess()) {
                 const Aws::Vector<Aws::Map<Aws::String, Aws::DynamoDB::Model::AttributeValue>> &items = result.GetResult().GetItems();
                 if (items.size() > 0) {
-                    std::cout << "Number of movies retrieved from scan: "
-                              << items.size() << std::endl;
-                    for (const auto &item: items) {
-                        printMovieInfo(item);
+                    std::stringstream stringStream;
+                    stringStream << "\nFound " << items.size() << " movies in one scan."
+                    << " How many would you like to see? ";
+                    size_t count = askQuestionForInt(stringStream.str());
+                     for (size_t i = 0; i < count && i < items.size(); ++i) {
+                        printMovieInfo(items[i]);
                     }
                 }
                 else {
-                    std::cout << "No item found in table: " << MOVIE_TABLE_NAME
-                              << std::endl;
+                    std::cout << "\nNo movies in the database between " << startYear <<
+                              " and " << endYear << "." << std::endl;
                 }
 
                 exclusiveStartKey = result.GetResult().GetLastEvaluatedKey();
+                if (exclusiveStartKey.size() > 0)
+                {
+                    std::cout << "Not all movies were retrieved. Scanning for more." << std::endl;
+                }
+                else{
+                    std::cout << "All movies were retrieved with this scan." << std::endl;
+                }
             }
             else {
                 std::cout << "Failed to Scan movies: "
@@ -409,20 +431,27 @@ bool AwsDoc::DynamoDB::dynamodbGettingStartedScenario(
 
     // 8. Delete a movie. (DeleteItem)
     {
-        Aws::DynamoDB::Model::DeleteItemRequest request;
-        request.AddKey(YEAR_KEY, Aws::DynamoDB::Model::AttributeValue().SetN(year));
-        request.AddKey(TITLE_KEY, Aws::DynamoDB::Model::AttributeValue().SetS(title));
-        request.SetTableName(MOVIE_TABLE_NAME);
+        std::stringstream stringStream;
+        stringStream << "\nWould you like to delete the movie " << title
+        << " from the database? (y/n) ";
+        Aws::String answer = askQuestion(stringStream.str());
+        if (answer == "y") {
+            Aws::DynamoDB::Model::DeleteItemRequest request;
+            request.AddKey(YEAR_KEY, Aws::DynamoDB::Model::AttributeValue().SetN(year));
+            request.AddKey(TITLE_KEY,
+                           Aws::DynamoDB::Model::AttributeValue().SetS(title));
+            request.SetTableName(MOVIE_TABLE_NAME);
 
-        const Aws::DynamoDB::Model::DeleteItemOutcome &result = dynamoClient.DeleteItem(
-                request);
-        if (result.IsSuccess()) {
-            std::cout << "The movie \"" << title << "\" was deleted." << std::endl;
-        }
-        else {
-            std::cout << "Failed to delete the  movie: "
-                      << result.GetError().GetMessage()
-                      << std::endl;
+            const Aws::DynamoDB::Model::DeleteItemOutcome &result = dynamoClient.DeleteItem(
+                    request);
+            if (result.IsSuccess()) {
+                std::cout << "\nRemoved \"" << title << "\" from the database." << std::endl;
+            }
+            else {
+                std::cerr << "Failed to delete the movie: "
+                          << result.GetError().GetMessage()
+                          << std::endl;
+            }
         }
     }
 
@@ -450,7 +479,6 @@ bool AwsDoc::DynamoDB::deleteDynamoTable(const Aws::String &tableName,
 
     return result.IsSuccess();
 }
-
 
 bool AwsDoc::DynamoDB::waitTableActive(const Aws::String &tableName,
                                        const Aws::Client::ClientConfiguration &clientConfiguration) {
@@ -579,6 +607,31 @@ float AwsDoc::DynamoDB::askQuestionForFloatRange(const Aws::String &string, floa
     return result;
 }
 
+int AwsDoc::DynamoDB::askQuestionForIntRange(const Aws::String &string, int low,
+                                                 int high) {
+    Aws::String resultString = askQuestion(string, [low, high](
+            const Aws::String &string1) -> bool {
+            try {
+                float number = std::atoi(string1.c_str());
+                return number >= low && number <= high;
+            }
+            catch (std::invalid_argument) {
+                return false;
+            }
+    });
+    float result = 0;
+    try {
+        result = std::atof(resultString.c_str());
+    }
+    catch (const std::invalid_argument&) {
+        std::cerr << "DynamoDB::askQuestionForFloatRange string not an int "
+                  << resultString << std::endl;
+    }
+
+    return result;
+}
+
+
 void AwsDoc::DynamoDB::printMovieInfo(
         const Aws::Map<Aws::String, Aws::DynamoDB::Model::AttributeValue> &movieMap) {
     {
@@ -633,8 +686,6 @@ int main(int argc, char **argv) {
 
 #endif // TESTING_BUILD
 
-static void zerr(int ret);
-
 Aws::String AwsDoc::DynamoDB::getMovieJSON() {
     const Aws::String JSON_FILE_NAME("moviedata.json");
     const Aws::String ZIP_FILE_NAME("movie.zip");
@@ -646,7 +697,8 @@ Aws::String AwsDoc::DynamoDB::getMovieJSON() {
                                                         "https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/samples/moviedata.zip"),
                                                 Aws::Http::HttpMethod::HTTP_GET,
                                                 Aws::Utils::Stream::DefaultResponseStreamFactoryMethod);
-    request->SetUserAgent("curl/7.79.1");
+   // request->SetUserAgent("curl/7.79.1");
+    std::cout << "Downloading the json file." << std::endl;
     auto response = httpClient->MakeRequest(request);
     Aws::String result;
 
@@ -654,12 +706,10 @@ Aws::String AwsDoc::DynamoDB::getMovieJSON() {
         FILE *src = fopen(ZIP_FILE_NAME.c_str(), "r");
         FILE *dst = fopen(JSON_FILE_NAME.c_str(), "w");
 
-        int zipResult = inflateZip(src, dst);
-        if (zipResult == Z_OK) {
-            std::cout << "Successful deflate" << std::endl;
-        }
-        else {
-            zerr(zipResult);
+        std::cout << "Unzipping the json file." << std::endl;
+        int zipResult = deflateZip(src, dst);
+        if (zipResult != Z_OK) {
+            std::cerr << "Could not deflate zip file" << std::endl;
         }
         fclose(src);
         fclose(dst);
@@ -686,7 +736,7 @@ Aws::String AwsDoc::DynamoDB::getMovieJSON() {
    invalid or incomplete, Z_VERSION_ERROR if the version of zlib.h and
    the version of the library linked do not match, or Z_ERRNO if there
    is an error reading or writing the files. */
-int AwsDoc::DynamoDB::inflateZip(FILE *source, FILE *dest) {
+int AwsDoc::DynamoDB::deflateZip(FILE *source, FILE *dest) {
     const int IN_CHUNK = 32767;
     const int OUT_CHUNK = 65535;
     int ret;
